+++ src/representations/msdr/msdrMusics.cpp
@@ -336,7 +336,7 @@ void msdrMusic::printShort (ostream& os) const
 
 #ifdef TRACING_IS_ENABLED
   // print the voice measures flat list
-  unsigned int voiceMeasuresFlatListSize =
+  size_t voiceMeasuresFlatListSize =
     fVoiceMeasuresFlatList.size ();
 
   os <<
